findCountry() updated, vars.py, last author added.
The findCountry() is now as optimized as possible to find and author's country within their affiliation. Vars.py created to hold lengthy dictionaries/lists used in the analyze.py script. Gender of the last author along with country added (regardless if they are last corresponding or not). The file name_gender_dataset.csv is needed to run analyze.py.